+++ src/main/java/kr/hhplus/be/server/domain/product/dto/ProductSearchResult.java
@@ -3,10 +3,12 @@ package kr.hhplus.be.server.domain.product.dto;
 import kr.hhplus.be.server.domain.product.Product;
 import kr.hhplus.be.server.domain.product.ProductStock;
 
+import java.math.BigDecimal;
+
 public record ProductSearchResult(
         Long id,
         String productName,
-        int productPrice,
+        BigDecimal productPrice,
         ProductStock stock
 ) {
